+++ changelog.md
@@ -1,3 +1,1 @@
-### Fabric
-- Added more informative icons
-- Fix menu open ability while not authenticated to webcam server
+- Add display_self_webcam server config field (default: false)